split out lightning module and pytorch model
