+++ configs.example/model/multimodal.yaml
@@ -1,106 +1,108 @@
-_target_: pvnet.models.multimodal.multimodal.Model
+_target_: pvnet.models.lightning_module.PVNetLightningModule
 
-output_quantiles: [0.02, 0.1, 0.25, 0.5, 0.75, 0.9, 0.98]
+model:
+  _target_: pvnet.models.multimodal.multimodal.Model
+  output_quantiles: [0.02, 0.1, 0.25, 0.5, 0.75, 0.9, 0.98]
 
-#--------------------------------------------
-# NWP encoder
-#--------------------------------------------
+  #--------------------------------------------
+  # NWP encoder
+  #--------------------------------------------
 
-nwp_encoders_dict:
-  ukv:
+  nwp_encoders_dict:
+    ukv:
+      _target_: pvnet.models.multimodal.encoders.encoders3d.DefaultPVNet
+      _partial_: True
+      in_channels: 2
+      out_features: 256
+      number_of_conv3d_layers: 6
+      conv3d_channels: 32
+      image_size_pixels: 24
+    ecmwf:
+      _target_: pvnet.models.multimodal.encoders.encoders3d.DefaultPVNet
+      _partial_: True
+      in_channels: 12
+      out_features: 256
+      number_of_conv3d_layers: 4
+      conv3d_channels: 32
+      image_size_pixels: 12
+
+  #--------------------------------------------
+  # Sat encoder settings
+  #--------------------------------------------
+
+  sat_encoder:
     _target_: pvnet.models.multimodal.encoders.encoders3d.DefaultPVNet
     _partial_: True
-    in_channels: 2
+    in_channels: 11
     out_features: 256
     number_of_conv3d_layers: 6
     conv3d_channels: 32
     image_size_pixels: 24
-  ecmwf:
-    _target_: pvnet.models.multimodal.encoders.encoders3d.DefaultPVNet
+
+  add_image_embedding_channel: False
+
+  #--------------------------------------------
+  # PV encoder settings
+  #--------------------------------------------
+
+  pv_encoder:
+    _target_: pvnet.models.multimodal.site_encoders.encoders.SingleAttentionNetwork
     _partial_: True
-    in_channels: 12
-    out_features: 256
-    number_of_conv3d_layers: 4
-    conv3d_channels: 32
-    image_size_pixels: 12
+    num_sites: 349
+    out_features: 40
+    num_heads: 4
+    kdim: 40
+    id_embed_dim: 20
 
-#--------------------------------------------
-# Sat encoder settings
-#--------------------------------------------
+  #--------------------------------------------
+  # Tabular network settings
+  #--------------------------------------------
 
-sat_encoder:
-  _target_: pvnet.models.multimodal.encoders.encoders3d.DefaultPVNet
-  _partial_: True
-  in_channels: 11
-  out_features: 256
-  number_of_conv3d_layers: 6
-  conv3d_channels: 32
-  image_size_pixels: 24
+  output_network:
+    _target_: pvnet.models.multimodal.linear_networks.networks.ResFCNet2
+    _partial_: True
+    fc_hidden_features: 128
+    n_res_blocks: 6
+    res_block_layers: 2
+    dropout_frac: 0.0
 
-add_image_embedding_channel: False
+  embedding_dim: 16
+  include_sun: True
+  include_gsp_yield_history: False
+  include_site_yield_history: False
 
-#--------------------------------------------
-# PV encoder settings
-#--------------------------------------------
+  # The mapping between the location IDs and their embedding indices
+  location_id_mapping:
+    1: 1
+    5: 2
+    110: 3
+  # ...
 
-pv_encoder:
-  _target_: pvnet.models.multimodal.site_encoders.encoders.SingleAttentionNetwork
-  _partial_: True
-  num_sites: 349
-  out_features: 40
-  num_heads: 4
-  kdim: 40
-  id_embed_dim: 20
+  #--------------------------------------------
+  # Times
+  #--------------------------------------------
 
-#--------------------------------------------
-# Tabular network settings
-#--------------------------------------------
+  # Foreast and time settings
+  forecast_minutes: 480
+  history_minutes: 120
 
-output_network:
-  _target_: pvnet.models.multimodal.linear_networks.networks.ResFCNet2
-  _partial_: True
-  fc_hidden_features: 128
-  n_res_blocks: 6
-  res_block_layers: 2
-  dropout_frac: 0.0
+  min_sat_delay_minutes: 60
 
-embedding_dim: 16
-include_sun: True
-include_gsp_yield_history: False
-include_site_yield_history: False
+  # These must also be set even if identical to forecast_minutes and  history_minutes
+  sat_history_minutes: 90
+  pv_history_minutes: 180
 
-# The mapping between the location IDs and their embedding indices
-location_id_mapping:
-  1: 1
-  5: 2
-  110: 3
-# ...
-
-#--------------------------------------------
-# Times
-#--------------------------------------------
-
-# Foreast and time settings
-forecast_minutes: 480
-history_minutes: 120
-
-min_sat_delay_minutes: 60
-
-# These must also be set even if identical to forecast_minutes and  history_minutes
-sat_history_minutes: 90
-pv_history_minutes: 180
-
-# These must be set for each NWP encoder
-nwp_history_minutes:
-  ukv: 120
-  ecmwf: 120
-nwp_forecast_minutes:
-  ukv: 480
-  ecmwf: 480
-# Optional; defaults to 60, so must be set for data with different time resolution
-nwp_interval_minutes:
-  ukv: 60
-  ecmwf: 60
+  # These must be set for each NWP encoder
+  nwp_history_minutes:
+    ukv: 120
+    ecmwf: 120
+  nwp_forecast_minutes:
+    ukv: 480
+    ecmwf: 480
+  # Optional; defaults to 60, so must be set for data with different time resolution
+  nwp_interval_minutes:
+    ukv: 60
+    ecmwf: 60
 
 # ----------------------------------------------
 # Optimizer
