+++ pvnet/models/base_model.py
@@ -2,28 +2,18 @@
 import copy
 import logging
 import os
-import tempfile
 import time
 from pathlib import Path
 
 import hydra
-import lightning.pytorch as pl
-import matplotlib.pyplot as plt
-import pandas as pd
 import pkg_resources
 import torch
-import torch.nn.functional as F
-import wandb
 import yaml
 from huggingface_hub import ModelCard, ModelCardData
 from huggingface_hub.file_download import hf_hub_download
 from huggingface_hub.hf_api import HfApi
-from ocf_data_sampler.torch_datasets.sample.base import copy_batch_to_device
 from torchvision.transforms.functional import center_crop
 
-from pvnet.models.utils import BatchAccumulator, MetricAccumulator, PredAccumulator
-from pvnet.optimizers import AbstractOptimizer
-from pvnet.utils import plot_batch_forecasts
 
 DATA_CONFIG_NAME = "data_config.yaml"
 MODEL_CONFIG_NAME = "model_config.yaml"
@@ -31,12 +21,6 @@ DATAMODULE_CONFIG_NAME = "datamodule_config.yaml"
 PYTORCH_WEIGHTS_NAME = "pytorch_model.bin"
 MODEL_CARD_NAME = "README.md"
 
-
-logger = logging.getLogger(__name__)
-
-activities = [torch.profiler.ProfilerActivity.CPU]
-if torch.cuda.is_available():
-    activities.append(torch.profiler.ProfilerActivity.CUDA)
 
 
 def make_clean_data_config(input_path, output_path, placeholder="PLACEHOLDER"):
@@ -144,33 +128,33 @@ def minimize_data_config(input_path, output_path, model):
 
 
 def download_hf_hub_with_retries(
-    repo_id,
-    filename,
-    revision,
-    cache_dir,
-    force_download,
-    proxies,
-    resume_download,
-    token,
-    local_files_only,
-    max_retries=5,
-    wait_time=10,
+    repo_id: str,
+    filename: str,
+    revision: str,
+    cache_dir: str | None,
+    force_download: bool,
+    proxies: dict | None,
+    resume_download: bool | None,
+    token: str | bool | None,
+    local_files_only: bool,
+    max_retries: int = 5,
+    wait_time: int = 10,
 ):
     """
     Tries to download a file from HuggingFace up to max_retries times.
 
     Args:
-        repo_id (str): HuggingFace repo ID
-        filename (str): Name of the file to download
-        revision (str): Specific model revision
-        cache_dir (str): Cache directory
-        force_download (bool): Whether to force a new download
-        proxies (dict): Proxy settings
-        resume_download (bool): Resume interrupted downloads
-        token (str): HuggingFace auth token
-        local_files_only (bool): Use local files only
-        max_retries (int): Maximum number of retry attempts
-        wait_time (int): Wait time (in seconds) before retrying
+        repo_id: HuggingFace repo ID
+        filename: Name of the file to download
+        revision: Specific model revision
+        cache_dir: Cache directory
+        force_download: Whether to force a new download
+        proxies: Proxy settings
+        resume_download: Resume interrupted downloads
+        token: HuggingFace auth token
+        local_files_only: Use local files only
+        max_retries: Maximum number of retry attempts
+        wait_time: Wait time (in seconds) before retrying
 
     Returns:
         str: The local file path of the downloaded file
@@ -202,10 +186,8 @@ def download_hf_hub_with_retries(
             time.sleep(wait_time)
 
 
-class PVNetModelHubMixin:
-    """
-    Implementation of [`PyTorchModelHubMixin`] to provide model Hub upload/download capabilities.
-    """
+class HuggingfaceMixin:
+    """Mixin for saving and loading model to and from huggingface"""
 
     @classmethod
     def from_pretrained(
@@ -319,8 +301,7 @@ class PVNetModelHubMixin:
         hf_repo_id: str | None = None,
         push_to_hub: bool = False,
     ) -> None:
-        """
-        Save weights in local directory.
+        """Save weights in local directory or upload to huggingface hub.
 
         Args:
             save_directory:
@@ -437,67 +418,39 @@ class PVNetModelHubMixin:
         )
 
 
-class BaseModel(pl.LightningModule, PVNetModelHubMixin):
+class BaseModel(torch.nn.Module, HuggingfaceMixin):
     """Abstract base class for PVNet submodels"""
 
     def __init__(
         self,
         history_minutes: int,
         forecast_minutes: int,
-        optimizer: AbstractOptimizer,
         output_quantiles: list[float] | None = None,
         target_key: str = "gsp",
         interval_minutes: int = 30,
-        timestep_intervals_to_plot: list[int] | None = None,
-        forecast_minutes_ignore: int = 0,
-        save_validation_results_csv: bool = False,
     ):
         """Abtstract base class for PVNet submodels.
 
         Args:
             history_minutes (int): Length of the GSP history period in minutes
             forecast_minutes (int): Length of the GSP forecast period in minutes
-            optimizer (AbstractOptimizer): Optimizer
             output_quantiles: A list of float (0.0, 1.0) quantiles to predict values for. If set to
                 None the output is a single value.
             target_key: The key of the target variable in the batch
             interval_minutes: The interval in minutes between each timestep in the data
-            timestep_intervals_to_plot: Intervals, in timesteps, to plot during training
-            forecast_minutes_ignore: Number of forecast minutes to ignore when calculating losses.
-                For example if set to 60, the model doesnt predict the first 60 minutes
-            save_validation_results_csv: whether to save full csv outputs from validation results.
         """
         super().__init__()
 
-        self._optimizer = optimizer
         self._target_key = target_key
-        if timestep_intervals_to_plot is not None:
-            for interval in timestep_intervals_to_plot:
-                assert type(interval) in [list, tuple] and len(interval) == 2, ValueError(
-                    f"timestep_intervals_to_plot must be a list of tuples or lists of length 2, "
-                    f"but got {timestep_intervals_to_plot=}"
-                )
-        self.time_step_intervals_to_plot = timestep_intervals_to_plot
-
-        # Model must have lr to allow tuning
-        # This setting is only used when lr is tuned with callback
-        self.lr = None
 
         self.history_minutes = history_minutes
         self.forecast_minutes = forecast_minutes
         self.output_quantiles = output_quantiles
         self.interval_minutes = interval_minutes
-        self.forecast_minutes_ignore = forecast_minutes_ignore
 
         # Number of timestemps for 30 minutely data
         self.history_len = history_minutes // interval_minutes
-        self.forecast_len = (forecast_minutes - forecast_minutes_ignore) // interval_minutes
-        self.forecast_len_ignore = forecast_minutes_ignore // interval_minutes
-
-        self._accumulated_metrics = MetricAccumulator()
-        self._accumulated_batches = BatchAccumulator(key_to_keep=self._target_key)
-        self._accumulated_y_hat = PredAccumulator()
-        self._horizon_maes = MetricAccumulator()
+        self.forecast_len = (forecast_minutes) // interval_minutes
 
         # Store whether the model should use quantile regression or simply predict the mean
         self.use_quantile_regression = self.output_quantiles is not None
@@ -507,10 +460,6 @@ class BaseModel(pl.LightningModule, PVNetModelHubMixin):
             self.num_output_features = self.forecast_len * len(self.output_quantiles)
         else:
             self.num_output_features = self.forecast_len
-
-        # save all validation results to array, so we can save these to weights n biases
-        self.validation_epoch_results = []
-        self.save_validation_results_csv = save_validation_results_csv
 
     def _adapt_batch(self, batch):
         """Slice batches into appropriate shapes for model.
@@ -575,10 +524,6 @@ class BaseModel(pl.LightningModule, PVNetModelHubMixin):
 
         return new_batch
 
-    def transfer_batch_to_device(self, batch, device, dataloader_idx):
-        """Method to move custom batches to a given device"""
-        return copy_batch_to_device(batch, device)
-
     def _quantiles_to_prediction(self, y_quantiles):
         """
         Convert network prediction into a point prediction.
@@ -596,364 +541,4 @@ class BaseModel(pl.LightningModule, PVNetModelHubMixin):
         """
         # y_quantiles Shape: batch_size, seq_length, num_quantiles
         idx = self.output_quantiles.index(0.5)
-        y_median = y_quantiles[..., idx]
-        return y_median
-
-    def _calculate_quantile_loss(self, y_quantiles, y):
-        """Calculate quantile loss.
-
-        Note:
-            Implementation copied from:
-                https://pytorch-forecasting.readthedocs.io/en/stable/_modules/pytorch_forecasting
-                /metrics/quantile.html#QuantileLoss.loss
-
-        Args:
-            y_quantiles: Quantile prediction of network
-            y: Target values
-
-        Returns:
-            Quantile loss
-        """
-        # calculate quantile loss
-        losses = []
-        for i, q in enumerate(self.output_quantiles):
-            errors = y - y_quantiles[..., i]
-            losses.append(torch.max((q - 1) * errors, q * errors).unsqueeze(-1))
-        losses = 2 * torch.cat(losses, dim=2)
-
-        return losses.mean()
-
-    def _calculate_common_losses(self, y, y_hat):
-        """Calculate losses common to train, and val"""
-
-        losses = {}
-
-        if self.use_quantile_regression:
-            losses["quantile_loss"] = self._calculate_quantile_loss(y_hat, y)
-            y_hat = self._quantiles_to_prediction(y_hat)
-
-        # calculate mse, mae
-        mse_loss = F.mse_loss(y_hat, y)
-        mae_loss = F.l1_loss(y_hat, y)
-
-        # TODO: Compute correlation coef using np.corrcoef(tensor with
-        # shape (2, num_timesteps))[0, 1] on each example, and taking
-        # the mean across the batch?
-        losses.update(
-            {
-                "MSE": mse_loss,
-                "MAE": mae_loss,
-            }
-        )
-
-        return losses
-
-    def _step_mae_and_mse(self, y, y_hat, dict_key_root):
-        """Calculate the MSE and MAE at each forecast step"""
-        losses = {}
-
-        mse_each_step = torch.mean((y_hat - y) ** 2, dim=0)
-        mae_each_step = torch.mean(torch.abs(y_hat - y), dim=0)
-
-        losses.update({f"MSE_{dict_key_root}/step_{i:03}": m for i, m in enumerate(mse_each_step)})
-        losses.update({f"MAE_{dict_key_root}/step_{i:03}": m for i, m in enumerate(mae_each_step)})
-
-        return losses
-
-    def _calculate_val_losses(self, y, y_hat):
-        """Calculate additional validation losses"""
-
-        losses = {}
-
-        if self.use_quantile_regression:
-            # Add fraction below each quantile for calibration
-            for i, quantile in enumerate(self.output_quantiles):
-                below_quant = y <= y_hat[..., i]
-                # Mask values small values, which are dominated by night
-                mask = y >= 0.01
-                losses[f"fraction_below_{quantile}_quantile"] = (below_quant[mask]).float().mean()
-
-            # Take median value for remaining metric calculations
-            y_hat = self._quantiles_to_prediction(y_hat)
-
-        # Log the loss at each time horizon
-        losses.update(self._step_mae_and_mse(y, y_hat, dict_key_root="horizon"))
-
-        # Log the persistance losses
-        y_persist = y[:, -1].unsqueeze(1).expand(-1, self.forecast_len)
-        losses["MAE_persistence/val"] = F.l1_loss(y_persist, y)
-        losses["MSE_persistence/val"] = F.mse_loss(y_persist, y)
-
-        # Log persistance loss at each time horizon
-        losses.update(self._step_mae_and_mse(y, y_persist, dict_key_root="persistence"))
-        return losses
-
-    def _training_accumulate_log(self, batch, batch_idx, losses, y_hat):
-        """Internal function to accumulate training batches and log results.
-
-        This is used when accummulating grad batches. Should make the variability in logged training
-        step metrics indpendent on whether we accumulate N batches of size B or just use a larger
-        batch size of N*B with no accumulaion.
-        """
-
-        losses = {k: v.detach().cpu() for k, v in losses.items()}
-        y_hat = y_hat.detach().cpu()
-
-        self._accumulated_metrics.append(losses)
-        self._accumulated_batches.append(batch)
-        self._accumulated_y_hat.append(y_hat)
-
-        if not self.trainer.fit_loop._should_accumulate():
-            losses = self._accumulated_metrics.flush()
-            batch = self._accumulated_batches.flush()
-            y_hat = self._accumulated_y_hat.flush()
-
-            self.log_dict(
-                losses,
-                on_step=True,
-                on_epoch=True,
-            )
-
-            # Number of accumulated grad batches
-            grad_batch_num = (batch_idx + 1) / self.trainer.accumulate_grad_batches
-
-            # We only create the figure every 8 log steps
-            # This was reduced as it was creating figures too often
-            if grad_batch_num % (8 * self.trainer.log_every_n_steps) == 0:
-                fig = plot_batch_forecasts(
-                    batch,
-                    y_hat,
-                    batch_idx,
-                    quantiles=self.output_quantiles,
-                    key_to_plot=self._target_key,
-                )
-                fig.savefig("latest_logged_train_batch.png")
-                plt.close(fig)
-
-    def training_step(self, batch, batch_idx):
-        """Run training step"""
-        y_hat = self(batch)
-
-        # Batch is adapted in the model forward method, but needs to be adapted here too
-        batch = self._adapt_batch(batch)
-
-        y = batch[self._target_key][:, -self.forecast_len :]
-
-        losses = self._calculate_common_losses(y, y_hat)
-        losses = {f"{k}/train": v for k, v in losses.items()}
-
-        self._training_accumulate_log(batch, batch_idx, losses, y_hat)
-
-        if self.use_quantile_regression:
-            opt_target = losses["quantile_loss/train"]
-        else:
-            opt_target = losses["MAE/train"]
-        return opt_target
-
-    def _log_forecast_plot(self, batch, y_hat, accum_batch_num, timesteps_to_plot, plot_suffix):
-        """Log forecast plot to wandb"""
-        fig = plot_batch_forecasts(
-            batch,
-            y_hat,
-            quantiles=self.output_quantiles,
-            key_to_plot=self._target_key,
-        )
-
-        plot_name = f"val_forecast_samples/batch_idx_{accum_batch_num}_{plot_suffix}"
-
-        try:
-            self.logger.experiment.log({plot_name: wandb.Image(fig)})
-        except Exception as e:
-            print(f"Failed to log {plot_name} to wandb")
-            print(e)
-        plt.close(fig)
-
-    def _log_validation_results(self, batch, y_hat, accum_batch_num):
-        """Append validation results to self.validation_epoch_results"""
-
-        # get truth values, shape (b, forecast_len)
-        y = batch[self._target_key][:, -self.forecast_len :]
-        y = y.detach().cpu().numpy()
-        batch_size = y.shape[0]
-
-        # get prediction values, shape (b, forecast_len, quantiles?)
-        y_hat = y_hat.detach().cpu().numpy()
-
-        # get time_utc, shape (b, forecast_len)
-        time_utc_key = f"{self._target_key}_time_utc"
-        time_utc = batch[time_utc_key][:, -self.forecast_len :].detach().cpu().numpy()
-
-        # get target id and change from (b,1) to (b,)
-        id_key = f"{self._target_key}_id"
-        target_id = batch[id_key].detach().cpu().numpy()
-        target_id = target_id.squeeze()
-
-        for i in range(batch_size):
-            y_i = y[i]
-            y_hat_i = y_hat[i]
-            time_utc_i = time_utc[i]
-            target_id_i = target_id[i]
-
-            results_dict = {
-                "y": y_i,
-                "time_utc": time_utc_i,
-            }
-            if self.use_quantile_regression:
-                results_dict.update(
-                    {f"y_quantile_{q}": y_hat_i[:, i] for i, q in enumerate(self.output_quantiles)}
-                )
-            else:
-                results_dict["y_hat"] = y_hat_i
-
-            results_df = pd.DataFrame(results_dict)
-            results_df["id"] = target_id_i
-            results_df["batch_idx"] = accum_batch_num
-            results_df["example_idx"] = i
-
-            self.validation_epoch_results.append(results_df)
-
-    def validation_step(self, batch: dict, batch_idx):
-        """Run validation step"""
-
-        accum_batch_num = batch_idx // self.trainer.accumulate_grad_batches
-
-        y_hat = self(batch)
-        # Batch is adapted in the model forward method, but needs to be adapted here too
-        batch = self._adapt_batch(batch)
-
-        y = batch[self._target_key][:, -self.forecast_len :]
-
-        if (batch_idx + 1) % self.trainer.accumulate_grad_batches == 0:
-            self._log_validation_results(batch, y_hat, accum_batch_num)
-
-        # Expand persistence to be the same shape as y
-        losses = self._calculate_common_losses(y, y_hat)
-        losses.update(self._calculate_val_losses(y, y_hat))
-
-        # Store these to make horizon accuracy plot
-        self._horizon_maes.append(
-            {i: losses[f"MAE_horizon/step_{i:03}"].cpu().numpy() for i in range(self.forecast_len)}
-        )
-
-        logged_losses = {f"{k}/val": v for k, v in losses.items()}
-
-        self.log_dict(
-            logged_losses,
-            on_step=False,
-            on_epoch=True,
-        )
-
-        # Make plots only if using wandb logger
-        if isinstance(self.logger, pl.loggers.WandbLogger) and accum_batch_num in [0, 1]:
-            # Store these temporarily under self
-            if not hasattr(self, "_val_y_hats"):
-                self._val_y_hats = PredAccumulator()
-                self._val_batches = BatchAccumulator(key_to_keep=self._target_key)
-
-            self._val_y_hats.append(y_hat)
-            self._val_batches.append(batch)
-
-            # if batch has accumulated
-            if (batch_idx + 1) % self.trainer.accumulate_grad_batches == 0:
-                y_hat = self._val_y_hats.flush()
-                batch = self._val_batches.flush()
-
-                self._log_forecast_plot(
-                    batch,
-                    y_hat,
-                    accum_batch_num,
-                    timesteps_to_plot=None,
-                    plot_suffix="all",
-                )
-
-                if self.time_step_intervals_to_plot is not None:
-                    for interval in self.time_step_intervals_to_plot:
-                        self._log_forecast_plot(
-                            batch,
-                            y_hat,
-                            accum_batch_num,
-                            timesteps_to_plot=interval,
-                            plot_suffix=f"timestep_{interval}",
-                        )
-
-                del self._val_y_hats
-                del self._val_batches
-
-        return logged_losses
-
-    def on_validation_epoch_end(self):
-        """Run on epoch end"""
-
-        try:
-            # join together validation results, and save to wandb
-            validation_results_df = pd.concat(self.validation_epoch_results)
-            validation_results_df["error"] = (
-                validation_results_df["y"] - validation_results_df["y_quantile_0.5"]
-            )
-
-            if isinstance(self.logger, pl.loggers.WandbLogger):
-                # log error distribution metrics
-                wandb.log(
-                    {
-                        "2nd_percentile_median_forecast_error": validation_results_df[
-                            "error"
-                        ].quantile(0.02),
-                        "5th_percentile_median_forecast_error": validation_results_df[
-                            "error"
-                        ].quantile(0.05),
-                        "95th_percentile_median_forecast_error": validation_results_df[
-                            "error"
-                        ].quantile(0.95),
-                        "98th_percentile_median_forecast_error": validation_results_df[
-                            "error"
-                        ].quantile(0.98),
-                        "95th_percentile_median_forecast_absolute_error": abs(
-                            validation_results_df["error"]
-                        ).quantile(0.95),
-                        "98th_percentile_median_forecast_absolute_error": abs(
-                            validation_results_df["error"]
-                        ).quantile(0.98),
-                    }
-                )
-            # saving validation result csvs
-            if self.save_validation_results_csv:
-                with tempfile.TemporaryDirectory() as tempdir:
-                    filename = os.path.join(tempdir, f"validation_results_{self.current_epoch}.csv")
-                    validation_results_df.to_csv(filename, index=False)
-
-                    # make and log wand artifact
-                    validation_artifact = wandb.Artifact(
-                        f"validation_results_epoch_{self.current_epoch}", type="dataset"
-                    )
-                    validation_artifact.add_file(filename)
-                    wandb.log_artifact(validation_artifact)
-
-        except Exception as e:
-            print("Failed to log validation results to wandb")
-            print(e)
-
-        self.validation_epoch_results = []
-        horizon_maes_dict = self._horizon_maes.flush()
-
-        # Create the horizon accuracy curve
-        if isinstance(self.logger, pl.loggers.WandbLogger):
-            per_step_losses = [[i, horizon_maes_dict[i]] for i in range(self.forecast_len)]
-            try:
-                table = wandb.Table(data=per_step_losses, columns=["horizon_step", "MAE"])
-                wandb.log(
-                    {
-                        "horizon_loss_curve": wandb.plot.line(
-                            table, "horizon_step", "MAE", title="Horizon loss curve"
-                        )
-                    },
-                )
-            except Exception as e:
-                print("Failed to log horizon_loss_curve to wandb")
-                print(e)
-
-    def configure_optimizers(self):
-        """Configure the optimizers using learning rate found with LR finder if used"""
-        if self.lr is not None:
-            # Use learning rate found by learning rate finder callback
-            self._optimizer.lr = self.lr
-        return self._optimizer(self)
+        return y_quantiles[..., idx]
